+++ src/api/source/SourceDescription.ts
@@ -55,6 +55,7 @@ export type CrossOriginSetting = '' | 'anonymous' | 'use-credentials';
  *
  * @category Source
  * @public
+ * @deprecated use {@link TypedSource.type} instead.
  */
 export enum SourceIntegrationId {
   THEO_LIVE = 'theolive'
@@ -306,6 +307,8 @@ export interface BaseSource {
    *
    * @remarks
    * <br/> - This can be used to signal that a source is specific to an integration.
+   *
+   * @deprecated use {@link TypedSource.type} instead.
    */
   integration?: SourceIntegrationId;
 
@@ -394,6 +397,8 @@ export interface TypedSource extends BaseSource {
    * <br/> - `'application/x-mpegURL'` or `'application/vnd.apple.mpegurl'`: The media resource is an HLS stream.
    * <br/> - `'video/mp4'`, `'video/webm'` and other formats: The media resource should use native HTML5 playback if supported by the browser.
    * <br/> - `'application/vnd.theo.hesp+json'`: The media resource is an HESP stream.
+   * <br/> - `'millicast'`: The media resource is a {@link MillicastSource | Millicast stream}.
+   * <br/> - `'theolive'`: The media resource is a {@link TheoLiveSource | THEOlive stream}.
    */
   type?: string;